pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "shortsyt"
version = "1.0.0"
description = "AI-powered autonomous YouTube Shorts generation and publishing pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Shortsyt Team"},
]
keywords = ["youtube", "shorts", "ai", "automation", "tts", "video-generation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "moviepy>=1.0.3",
    "edge-tts>=6.1",
    "pysrt>=1.1.2",
    "yt-dlp>=2024.1",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "pandas>=2.0",
    "opencv-python>=4.8",
    "tqdm>=4.65",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.3",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
